+++ index.html
@@ -162,6 +162,6 @@
     <p>Todos os direitos reservados</p>
   </footer>
 
-  <script src="script.js"></script>
+  <script type="module" src="js/script.js"></script>
 </body>
 </html>
