+++ js/script.js
@@ -1,0 +1,9 @@
+import initScroll from "./modules/scroll.js";
+import initAnimeScroll from "./modules/anime-scroll.js";
+import initAccordion from "./modules/accordion.js"
+import initTabNav from "./modules/tab-nav.js";
+
+initScroll();
+initAnimeScroll();
+initAccordion();
+initTabNav();
