--- script.js
@@ -1,121 +1,0 @@
-// função de navegação da lista de animais
-function initTabNav() {
-  //seleciona as classes tabMenu e tabContent da lista de animais e sections da lista de descrições(1)
-  const tabMenu = document.querySelectorAll('[data-tab="menu"] li');
-  const tabContent = document.querySelectorAll('[data-tab="content"] section');
-
-  //TESTE(5)
-  //se houver itens na tabMenu e/ou tabContent o código vai acontecer
-  if(tabMenu.length && tabContent.length) {
-
-    // faz com que o primeiro item da section sempre fique ativa ao iniciar a página, monstrando assim seu conteúdo mesmo não tendo sido clicado(4)
-    tabContent[0].classList.add('active');
-
-    //função que ativa e desativa as sections, adicionando e removendo a classe 'active'(2)
-    function activeTab(index) {
-      tabContent.forEach((section) => {
-        section.classList.remove('active');
-      });
-      const direction = tabContent[index].dataset.anime;
-      tabContent[index].classList.add('active', direction);
-    }
-
-    //loop em que, para cada item da lista de animais, ao acontecer o evento de 'click', será adicionado a class 'active' a cada um dos itens que forem ativados(3)
-    tabMenu.forEach((itemMenu, index) => {
-      itemMenu.addEventListener('click', () => {
-        activeTab(index);
-      });
-    });
-  }
-}
-initTabNav();
-
-
-// função da lista do FAQ
-function initAccordion() {
-  //seleciona os tópicos do FAQ(1)
-  const accordionList = document.querySelectorAll('[data-anime="accordion"] dt');
-  //constante de ativação
-  const activeClass = 'active';
-
-  // TESTE(5)
-  // Se houver itens no 'accordionList' a função será ativada
-  if(accordionList.length) {
-    //faz com que o primeiro tópico e descrição já fique ativo ao iniciar o código(4)
-    accordionList[0].classList.add(activeClass);
-    accordionList[0].nextElementSibling.classList.add(activeClass);
-
-    //função que adiciona a classe 'active' aos tópicos e descrições no evento de 'click'(3)
-    function activeAccordion() {
-      this.classList.toggle(activeClass);
-      this.nextElementSibling.classList.toggle(activeClass);
-    }
-
-    //adiciona evento de 'click',(2)
-    accordionList.forEach((item) => {
-      item.addEventListener('click', activeAccordion);
-    });
-  }
-}
-initAccordion();
-
-
-//função do scroll suave
-function initScroll() {
-  //seleciona todos os links internos(1)
-  const innerLinks = document.querySelectorAll('[data-menu="suave"] a[href^="#"]');
-
-  //função que linka o item do navegador com a sua seção(3)
-  function scrollToSection(event) {
-    event.preventDefault();
-    const href = event.currentTarget.getAttribute('href'); //pega o atributo do href
-    const section = document.querySelector(href);
-    
-    //faz o movimento de scroll do item para a seção(4)
-    section.scrollIntoView({
-      behavior: 'smooth',
-      block: 'start',
-    });
-  }
-
-  //loop que adiciona para cada link um evento de 'click'(2)
-  innerLinks.forEach((link) => {
-    link.addEventListener('click', scrollToSection);
-  });
-}
-initScroll();
-
-
-//função de animação das seções durante o scroll
-function initAnimeScroll() {
-  //seleciona as sections com 'js-scroll'(1)
-  const sections = document.querySelectorAll('[data-anime="scroll"]');
-
-  // TESTE(5)
-  //se houver itens na sections, o código vai funcionar
-  if(sections.length) {
-
-    //calcula o tamanho de 60% da tela
-    const halfWindow = window.innerHeight * 0.6;
-
-    //função que dispara o scroll(2)
-    function animeScroll() {
-      sections.forEach((section) => { 
-        const sectionTop = section.getBoundingClientRect().top; //pega a distância do topo de cada elemento
-        const isSectionVisible = (sectionTop - halfWindow) < 0; //calcula se a seção está visível na tela do browser, se for true, adiciona a classe 'active', se não, remove a classe 'active'
-        if(isSectionVisible) { 
-          section.classList.add('active');
-        } else {
-          section.classList.remove('active');
-        }
-      })
-    }
-
-    //deixa a primeira seção ativada (4)
-    animeScroll();
-
-    //adiciona o evento no Scroll(3)
-    window.addEventListener('scroll', animeScroll);
-  }
-}
-initAnimeScroll();
